# apps/api/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "Confidence": {
        "properties": {
          "level": {
            "$ref": "#/components/schemas/ConfidenceLevel"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "level"
        ],
        "title": "Confidence",
        "type": "object"
      },
      "ConfidenceLevel": {
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "ConfidenceLevel",
        "type": "string"
      },
      "CuratedLocation": {
        "properties": {
          "conditions_region": {
            "title": "Conditions Region",
            "type": "string"
          },
          "coops_station": {
            "title": "Coops Station",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ndbc_stations": {
            "items": {
              "type": "string"
            },
            "title": "Ndbc Stations",
            "type": "array"
          },
          "nws_zone": {
            "title": "Nws Zone",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "temp_offset": {
            "default": 0,
            "title": "Temp Offset",
            "type": "integer"
          },
          "temp_region": {
            "title": "Temp Region",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "state",
          "lat",
          "lng",
          "timezone",
          "coops_station",
          "ndbc_stations",
          "nws_zone",
          "temp_region",
          "conditions_region"
        ],
        "title": "CuratedLocation",
        "type": "object"
      },
      "DisplayUnits": {
        "enum": [
          "imperial",
          "metric"
        ],
        "title": "DisplayUnits",
        "type": "string"
      },
...
      "ForecastEnvelope": {
        "description": "The top-level response shape named in docs/CANONICAL_ROADMAP.md's\n\"Required API surface\" section. `conditions`, `tides`, `hourly_outlook`,\nand `recommendations` are opaque here on purpose \u2014 see module docstring.",
        "properties": {
          "conditions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "hourly_outlook": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hourly Outlook"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "recommendations": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendations"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "title": "Sources",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/ForecastState"
          },
          "tides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tides"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "location",
          "generated_at",
          "state",
          "confidence"
        ],
        "title": "ForecastEnvelope",
        "type": "object"
      },
      "ForecastState": {
        "description": "docs/product-definition.md \u00a7 Product vocabulary \u00a7 Forecast state.",
        "enum": [
          "fresh",
          "stale",
          "partial",
          "unavailable"
        ],
        "title": "ForecastState",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Location": {
        "description": "A resolved location. Coordinates are rounded before storage/transit\nper docs/CANONICAL_ROADMAP.md's data-minimization requirement \u2014 rounding\nitself happens at resolution time (sprint 18/19), not here.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "lat": {
            "description": "Rounded, not precise \u2014 see module docstring.",
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "description": "Rounded, not precise \u2014 see module docstring.",
            "title": "Lng",
            "type": "number"
          },
          "marine_zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marine Zone"
          },
          "station_ids": {
            "items": {
              "type": "string"
            },
            "title": "Station Ids",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "lat",
          "lng",
          "timezone"
        ],
        "title": "Location",
        "type": "object"
      },
      "LocationResolveRequest": {
        "description": "Exactly one of `location_id`, or both `lat` and `lng`, must be\ngiven \u2014 never both shapes, never neither.",
        "properties": {
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          }
        },
        "title": "LocationResolveRequest",
        "type": "object"
      },
      "PreferencesOut": {
        "description": "Response shape. Always returns a full object, even for a user with\nno row yet (sprint 36's endpoint creates one with these same\ndefaults on first read) -- a `404` for \"no preferences set yet\" would\nmake every new user's first `GET` an error case for no real reason.",
        "properties": {
          "default_location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Location Id"
          },
          "fishing_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fishing Style"
          },
          "surf_threshold_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surf Threshold Ft"
          },
          "units": {
            "$ref": "#/components/schemas/DisplayUnits",
            "default": "imperial"
          },
          "wind_threshold_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Threshold Kt"
          }
        },
        "title": "PreferencesOut",
        "type": "object"
      },
      "PreferencesUpdate": {
        "description": "`PATCH` body -- every field optional and omitted fields are left\nunchanged (a real partial update, not a `PUT`-style full replace).\n`None` is a meaningful, distinct value from \"omitted\": it clears a\npreviously-set threshold/style/default location, which is why every\nfield below is `T | None` rather than plain `T` with an implicit\ncan't-unset gap. `exclude_unset=True` at the call site is what\nreads that distinction back out of the request body.",
        "properties": {
          "default_location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Location Id"
          },
          "fishing_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fishing Style"
          },
          "surf_threshold_ft": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surf Threshold Ft"
          },
          "units": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DisplayUnits"
              },
              {
                "type": "null"
              }
            ]
          },
          "wind_threshold_kt": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Threshold Kt"
          }
        },
        "title": "PreferencesUpdate",
        "type": "object"
      },
      "ResolvedLocation": {
        "description": "A forecast-ready location \u2014 either a curated spot as-is, or an\nassembled dynamic point. `is_dynamic=False` means every field came\ndirectly from `CuratedLocation`; `is_dynamic=True` means station IDs\nwere resolved from live catalogs and the coarse regional fields were\ninherited from the nearest curated neighbor (see\n`resolve_dynamic_location`).\n\n`anchor_miles` is `None` for a curated location \u2014 it *is* the named\nspot, so there's no meaningful \"distance to the nearest real\nstation\" to report \u2014 and for a dynamic point where no anchor\n(curated neighbor, CO-OPS station, or NDBC buoy) was found at all.\nOtherwise it's `resolve_dynamic_location`'s nearest-anchor distance,\nadded so a later confidence assessment (`app.domain.confidence`) has\nreal distance data to work with instead of none at all.",
        "properties": {
          "anchor_miles": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Miles"
          },
          "conditions_region": {
            "title": "Conditions Region",
            "type": "string"
          },
          "coops_station": {
            "title": "Coops Station",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_dynamic": {
            "title": "Is Dynamic",
            "type": "boolean"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ndbc_stations": {
            "items": {
              "type": "string"
            },
            "title": "Ndbc Stations",
            "type": "array"
          },
          "nws_zone": {
            "title": "Nws Zone",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "temp_offset": {
            "title": "Temp Offset",
            "type": "integer"
          },
          "temp_region": {
            "title": "Temp Region",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "water_temp_station": {
            "title": "Water Temp Station",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "state",
          "lat",
          "lng",
          "timezone",
          "coops_station",
          "water_temp_station",
          "ndbc_stations",
          "nws_zone",
          "temp_region",
          "conditions_region",
          "temp_offset",
          "is_dynamic"
        ],
        "title": "ResolvedLocation",
        "type": "object"
      },
      "SaveLocationRequest": {
        "properties": {
          "location_id": {
            "title": "Location Id",
            "type": "string"
          }
        },
        "required": [
          "location_id"
        ],
        "title": "SaveLocationRequest",
        "type": "object"
      },
      "SavedLocationOut": {
        "description": "`location_id`'s display fields (`name`/`state`/`lat`/`lng`) are\nresolved fresh at read time via the same `resolve_location_id` every\nforecast lookup uses, not stored and left to go stale -- a curated\nlocation's own name/state never changes, but resolving live keeps\nthis one code path instead of two ways to know a location's name.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "location_id": {
            "title": "Location Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "saved_at": {
            "format": "date-time",
            "title": "Saved At",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "location_id",
          "name",
          "state",
          "lat",
          "lng",
          "saved_at"
        ],
        "title": "SavedLocationOut",
        "type": "object"
      },
      "SourceState": {
        "enum": [
          "ok",
          "degraded",
          "unavailable"
        ],
        "title": "SourceState",
        "type": "string"
      },
      "SourceStatus": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/SourceState"
          }
        },
        "required": [
          "provider",
          "state",
          "as_of"
        ],
        "title": "SourceStatus",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Warning": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/WarningSeverity"
          }
        },
        "required": [
          "code",
          "message",
          "severity"
        ],
        "title": "Warning",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Saltline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/forecasts/{location_id}": {
      "get": {
        "operationId": "get_forecast_v1_forecasts__location_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Forecast",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/v1/forecasts/{location_id}/refresh": {
      "post": {
        "operationId": "refresh_forecast_v1_forecasts__location_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Forecast",
        "tags": [
          "forecasts"
        ]
      }
    },
...
    "/v1/locations/resolve": {
      "post": {
        "operationId": "resolve_location_v1_locations_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedLocation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Location",
        "tags": [
          "locations"
        ]
      }
    },
    "/v1/locations/search": {
      "get": {
        "operationId": "search_locations_v1_locations_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CuratedLocation"
                  },
                  "title": "Response Search Locations V1 Locations Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Locations",
        "tags": [
          "locations"
        ]
      }
    },
...
    "/v1/me/locations": {
      "get": {
        "operationId": "list_saved_locations_v1_me_locations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedLocationOut"
                  },
                  "title": "Response List Saved Locations V1 Me Locations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Saved Locations",
        "tags": [
          "saved-locations"
        ]
      },
      "post": {
        "operationId": "save_location_v1_me_locations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveLocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedLocationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Location",
        "tags": [
          "saved-locations"
        ]
      }
    },
    "/v1/me/locations/{saved_location_id}": {
      "delete": {
        "operationId": "delete_saved_location_v1_me_locations__saved_location_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "saved_location_id",
            "required": true,
            "schema": {
              "title": "Saved Location Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Saved Location",
        "tags": [
          "saved-locations"
        ]
      }
    },
    "/v1/me/preferences": {
      "get": {
        "operationId": "get_preferences_v1_me_preferences_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Preferences",
        "tags": [
          "preferences"
        ]
      },
      "patch": {
        "operationId": "update_preferences_v1_me_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Preferences",
        "tags": [
          "preferences"
        ]
      }
    }
  }
}